% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{convert_units()}
\usage{
convert_units(dat, to_units)
}
\arguments{
\item{dat}{list of datasets CDMO datasets}

\item{to_units}{unit set to convert from default SWMPrStorm units ('English' or 'SI')}
}
\value{
df, list of datasets in the same input format as dat
}
\description{
convert_units()
}
\examples{
\dontrun{
convert_units(dat, 'English')
}
}
