% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data_table.R
\name{daily_data_table}
\alias{daily_data_table}
\title{daily_data_table}
\usage{
daily_data_table(
  var_in,
  data_path,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  reserve = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  skip = NULL,
  ...
)
}
\arguments{
\item{skip}{}
}
\value{

}
\description{
daily_data_table
}
