% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_roc.R
\name{event_roc}
\alias{event_roc}
\title{Title}
\usage{
event_roc(
  var_in,
  data_path = NULL,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
