% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries_dual.R
\name{event_timeseires_dual}
\alias{event_timeseires_dual}
\title{Title}
\usage{
event_timeseires_dual(
  var_in,
  data_path,
  storm_nm = NULL,
  view_start = NULL,
  view_end = NULL,
  stn_wq = NULL,
  stn_met = NULL,
  param_primary = NULL,
  param_secondary = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
