% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries.R
\name{event_timeseries}
\alias{event_timeseries}
\title{event_timeseries}
\usage{
event_timeseries(
  var_in,
  data_path = NULL,
  reserve = NULL,
  storm_start = NULL,
  storm_end = NULL,
  view_start = NULL,
  view_end = NULL,
  recovery_start = NULL,
  recovery_end = NULL,
  stn_wq = NULL,
  keep_flags = NULL,
  skip = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{reserve}{3 digit reserve code (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS start of storm event (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS end of storm event (string).}

\item{view_start}{YYYY-MM-DD HH:MM:SS first datetime of data to plot (string).}

\item{view_end}{YYYY-MM-DD HH:MM:SS last datetime of data to plot (string).}

\item{recovery_start}{YYYY-MM-DD HH:MM:SS start of storm recovery period (string).}

\item{recovery_end}{YYYY-MM-DD HH:MM:SS end of storm recovery period (string).}

\item{stn_wq}{station code to analyze. if blank all active stations within reserve will be analyzed (string)}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}
}
\value{

}
\description{
event_timeseries
}
