% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries_hourly.R
\name{event_timeseries_hourly}
\alias{event_timeseries_hourly}
\title{event_timeseries_hourly}
\usage{
event_timeseries_hourly(var_in)
}
\arguments{
\item{var_in}{}
}
\value{

}
\description{
event_timeseries_hourly
}
