% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries_precip.R
\name{event_timeseries_precip}
\alias{event_timeseries_precip}
\title{event_timeseries_precip}
\usage{
event_timeseries_precip(
  var_in,
  data_path,
  reserve = NULL,
  storm_start = NULL,
  storm_end = NULL,
  stn_met = NULL,
  keep_flags = NULL,
  flip = FALSE,
  skip = NULL,
  ...
)
}
\arguments{
\item{skip}{}
}
\value{

}
\description{
event_timeseries_precip
}
