% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_windrose.R
\name{event_windrose}
\alias{event_windrose}
\title{event_windrose}
\usage{
event_windrose(
  var_in,
  data_path = NULL,
  reserve = NULL,
  storm_start = NULL,
  storm_end = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  skip = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{reserve}{3 digit reserve code (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS end of storm event (string).}

\item{met_sites}{comma separated list of station codes (string)}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}
}
\value{

}
\description{
event_windrose
}
