% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{initialize_project()}
\usage{
initialize_project(path_to_project = NULL)
}
\arguments{
\item{path_to_project}{pathway to project location. If blank, the current working directory will be called. (string).}
}
\value{
no objects returned. standard SWMPrStorm empty output directories are created.
}
\description{
initialize_project()
}
\examples{

\dontrun{
## initialize in current working director:
initialize_project()
## initialize in specific location
working_dir <- 'C:/path/to/directory'
initialize_project(working_dir)
}
}
