% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_lab_dir_stn_order.R
\name{print_lab_dir_stn_order}
\alias{print_lab_dir_stn_order}
\title{print_lab_dir_stn_order()}
\usage{
print_lab_dir_stn_order(reserve)
}
\arguments{
\item{reserve}{3-digit reserve code}
}
\description{
print_lab_dir_stn_order()
}
\examples{
\dontrun{
print_lab_dir_stn_order('del')
}
}
