% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_storm_track.R
\name{single_storm_track}
\alias{single_storm_track}
\title{Title}
\usage{
single_storm_track(
  map_in,
  nerr_site_id = NULL,
  storm_nm = NULL,
  storm_rank = NULL,
  stations = NULL,
  bbox = NULL,
  path_to_shp = NULL,
  path_to_base = NULL,
  lab_loc = NULL,
  scale_pos = "bottomleft"
)
}
\arguments{
\item{scale_pos}{}
}
\value{

}
\description{
Title
}
