% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_param_check.R
\name{std_param_check}
\alias{std_param_check}
\title{Check user parameter against standard parameter list}
\usage{
std_param_check(param)
}
\arguments{
\item{param}{}
}
\value{

}
\description{
Check user parameter against standard parameter list
}
