% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data_table.R
\name{summary_data_table}
\alias{summary_data_table}
\title{summary_data_table}
\usage{
summary_data_table(
  var_in,
  data_path,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  reserve = NULL,
  keep_flags = NULL,
  skip = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{storm_nm}{name of storm event (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS (string).}

\item{reserve}{3 digit reserve code (string).}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}
}
\value{

}
\description{
summary_data_table
}
