% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_table.R
\docType{data}
\name{unit_table}
\alias{unit_table}
\title{Unit Definitions for NERRS parameter data}
\format{
A \code{\link[base]{data.frame}} object
}
\usage{
data(unit_table)
}
\description{
Formatted for use with SWMPrStorm Package.
Default units are default to the SWMPrStorm analysis, not the CDMO export.
}
\keyword{datasets}
