% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_axis_unit_labeler.R
\name{y_axis_unit_labeler}
\alias{y_axis_unit_labeler}
\title{Generate y-axis Label Based on SWMP Parameter Abbreviation}
\usage{
y_axis_unit_labeler(param, user_units)
}
\arguments{
\item{param}{chr string of variable abbreviation}

\item{user_units}{chr string equal to 'English" or 'SI"}
}
\value{
Returns character vector that can be used as a y-axis label in ggplot2. includes special characters.
}
\description{
Generate y-axis Label Based on SWMP Parameter Abbreviation
}
\details{
A helper function used internally by several plotting functions to generate y-axis labels. This function does not convert sample results from metric to english. It only adjusts the units in the y-axis label.
}
\examples{
\dontrun{
y_axis_unit_labeler <- y_labeler('do_mgl', 'English')
}

}
