% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_axis_unit_labeler_delta.R
\name{y_axis_unit_labeler_delta}
\alias{y_axis_unit_labeler_delta}
\title{y_axis_unit_labeler_delta}
\usage{
y_axis_unit_labeler_delta(param, user_units)
}
\arguments{
\item{param}{wq or met cdmo parameter code}

\item{user_units}{'English' or 'SI'}
}
\value{
returns a chr string for y-axis labels in roc plots with specialized characters
}
\description{
y_axis_unit_labeler_delta
}
\examples{

\dontrun{
parameter <- 'temp'
units <- 'English'
y_axis_unit_labeler_delta(parameter, units)
}
}
