% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_one_reserve_one_event.R
\name{compare_one_reserve_one_event}
\alias{compare_one_reserve_one_event}
\title{produce daily stat time series for single event at one reserve}
\usage{
compare_one_reserve_one_event(
  var_in,
  data_path,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  reserve = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
produce daily stat time series for single event at one reserve
}
