% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data_table.R
\name{daily_data_table}
\alias{daily_data_table}
\title{daily_data_table}
\usage{
daily_data_table(
  var_in,
  data_path = NULL,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  reserve = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  skip = NULL,
  user_units = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{storm_nm}{name of storm event (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS start of storm event (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS end of storm event (string).}

\item{reserve}{3 digit reserve code (string).}

\item{wq_sites}{comma separated list of wq stations to plot. if left blank, all active stations from reserve will be plotted. (string).}

\item{met_sites}{comma separated list of met stations to plot. if left blank, all active stations from reserve will be plotted. (string).}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}

\item{user_units}{User defined units. Set to "English" or "SI". Default CDMO data is in SI units. Not all parameters have common equivalent English units (e.g. concentrations), and therefore not all will be converted.}
}
\value{
tables are generated and saved in /output/wq/data_table/ and /output/met/data_table/
}
\description{
daily_data_table
}
\examples{

\dontrun{
#StormVariables.xlsx is a template variable input file saved in data/
vars_in <- 'data/StormTrackVariables.xlsx'
single_storm_track(var_in = vars_in)
}
}
