% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_roc.R
\name{event_roc}
\alias{event_roc}
\title{event_roc}
\usage{
event_roc(
  var_in,
  data_path = NULL,
  reserve = NULL,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  skip = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{reserve}{3 digit reserve code (string).}

\item{storm_nm}{name of storm event (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS start of storm event (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS end of storm event (string).}

\item{wq_sites}{comma separated list of wq stations to plot. if left blank, all active stations from reserve will be plotted. (string).}

\item{met_sites}{comma separated list of met stations to plot. if left blank, all active stations from reserve will be plotted. (string).}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}
}
\value{

}
\description{
event_roc
}
