% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries.R
\name{event_timeseries}
\alias{event_timeseries}
\title{event_timeseries}
\usage{
event_timeseries(
  var_in,
  data_path = NULL,
  storm_start = NULL,
  storm_end = NULL,
  view_start = NULL,
  view_end = NULL,
  recovery_start = NULL,
  recovery_end = NULL,
  stn_wq = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
event_timeseries
}
