% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries_hourly.R
\name{event_timeseries_hourly}
\alias{event_timeseries_hourly}
\title{event_timeseries_hourly}
\usage{
event_timeseries_hourly(
  var_in,
  data_path = NULL,
  reserve = NULL,
  storm_nm = NULL,
  onset_start = NULL,
  onset_end = NULL,
  view_start = NULL,
  view_end = NULL,
  recovery_start = NULL,
  recovery_end = NULL,
  wq_sites = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{reserve}{}
}
\value{

}
\description{
event_timeseries_hourly
}
