% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_timeseries_precip.R
\name{event_timeseries_precip}
\alias{event_timeseries_precip}
\title{event_timeseries_precip}
\usage{
event_timeseries_precip(
  var_in,
  data_path,
  reserve = NULL,
  storm_start = NULL,
  storm_end = NULL,
  stn_met = NULL,
  keep_flags = NULL,
  flip = FALSE,
  skip = NULL
)
}
\arguments{
\item{var_in}{.xlsx with all required input variables defined (string).}

\item{data_path}{pathway to cdmo data folder (string).}

\item{reserve}{3 digit reserve code (string).}

\item{storm_start}{YYYY-MM-DD HH:MM:SS start of storm event (string).}

\item{storm_end}{YYYY-MM-DD HH:MM:SS end of storm event (string).}

\item{stn_met}{comma separated list of data quality flags that should be kept (string).}

\item{keep_flags}{comma separated list of data quality flags that should be kept (string).}

\item{flip}{TRUE/FALSE. If TRUE, the plot axes will flip.}

\item{skip}{TRUE/FALSE. If TRUE, function will be skipped (string).}
}
\value{

}
\description{
event_timeseries_precip
}
