% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_windrose.R
\name{event_windrose}
\alias{event_windrose}
\title{event_windrose}
\usage{
event_windrose(
  var_in,
  data_path = NULL,
  reserve = NULL,
  storm_start = NULL,
  storm_end = NULL,
  met_sites = NULL,
  keep_flags = NULL,
  ...
)
}
\arguments{
\item{reserve}{}
}
\value{

}
\description{
event_windrose
}
