% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project}
\alias{initialize_project}
\title{initialize_project()}
\usage{
initialize_project(path_to_project = NULL)
}
\arguments{
\item{path_to_project}{}
}
\value{

}
\description{
initialize_project()
}
