% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_storm_track.R
\name{multi_storm_track}
\alias{multi_storm_track}
\title{multi_storm_track}
\usage{
multi_storm_track(
  map_in,
  nerr_site_id = NULL,
  storm_nm = NULL,
  storm_rank = NULL,
  stations = NULL,
  bbox = NULL,
  path_to_shp = NULL,
  path_to_base = NULL,
  lab_loc = NULL,
  scale_pos = "bottomleft",
  skip = NULL
)
}
\arguments{
\item{skip}{}
}
\value{

}
\description{
multi_storm_track
}
