% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_storm_track.R
\name{res_storm_track}
\alias{res_storm_track}
\title{res_storm_track}
\usage{
res_storm_track(
  nerr_site_id = NULL,
  storm_nm = NULL,
  stations = NULL,
  bbox = NULL,
  path_to_shp = NULL,
  lab_loc = NULL,
  scale_pos = "bottomleft"
)
}
\arguments{
\item{scale_pos}{}
}
\value{

}
\description{
res_storm_track
}
