% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_data_table.R
\name{summary_data_table}
\alias{summary_data_table}
\title{summary_data_table}
\usage{
summary_data_table(
  var_in,
  data_path,
  storm_nm = NULL,
  storm_start = NULL,
  storm_end = NULL,
  reserve = NULL,
  keep_flags = NULL,
  skip = NULL
)
}
\arguments{
\item{skip}{}
}
\value{

}
\description{
summary_data_table
}
